+++ src/Components/Admin/Content/ModalCreateUser.js
@@ -61,6 +61,7 @@ function ModalCreateUser(props) {
          if (data && data.EC === 0) {
             toast.success(data.EM);
             handleClose();
+            await props.fetchAllUser();
          }
          if (data && data.EC !== 0) {
             toast.error(data.EM);
